+++ ggpo/gui/ui/ggpowindow_ui.py
@@ -2,7 +2,7 @@
 
 # Form implementation generated from reading ui file 'ggpo/gui/ui/ggpowindow.ui'
 #
-# Created: Sat Mar 22 14:43:05 2014
+# Created: Sat Mar 22 15:59:28 2014
 #      by: PyQt4 UI code generator 4.9.1
 #
 # WARNING! All changes made in this file will be lost!
@@ -150,6 +150,8 @@ class Ui_MainWindow(object):
         self.uiNormalThemeAct.setObjectName(_fromUtf8("uiNormalThemeAct"))
         self.action0 = QtGui.QAction(MainWindow)
         self.action0.setObjectName(_fromUtf8("action0"))
+        self.uiCustomEmoticonsAct = QtGui.QAction(MainWindow)
+        self.uiCustomEmoticonsAct.setObjectName(_fromUtf8("uiCustomEmoticonsAct"))
         self.menuAction.addAction(self.uiAwayAct)
         self.menuAction.addAction(self.uiFocusOnChatAct)
         self.menuAction.addAction(self.uiEmoticonAct)
@@ -170,6 +172,7 @@ class Ui_MainWindow(object):
         self.menuSetting.addAction(self.uiFontAct)
         self.menuSetting.addAction(self.uiThemeMenu.menuAction())
         self.menuSetting.addAction(self.uiSmoothingMenu.menuAction())
+        self.menuSetting.addAction(self.uiCustomEmoticonsAct)
         self.menuSetting.addSeparator()
         self.menuSetting.addAction(self.uiLocateGgpofbaAct)
         self.menuSetting.addAction(self.uiLocateWineAct)
@@ -254,5 +257,6 @@ class Ui_MainWindow(object):
         self.uiCustomQssFileAct.setText(QtGui.QApplication.translate("MainWindow", "Custom Qss File", None, QtGui.QApplication.UnicodeUTF8))
         self.uiNormalThemeAct.setText(QtGui.QApplication.translate("MainWindow", "&Normal", None, QtGui.QApplication.UnicodeUTF8))
         self.action0.setText(QtGui.QApplication.translate("MainWindow", "0", None, QtGui.QApplication.UnicodeUTF8))
+        self.uiCustomEmoticonsAct.setText(QtGui.QApplication.translate("MainWindow", "Custom &Emoticons", None, QtGui.QApplication.UnicodeUTF8))
 
 from ggpo.gui.completionlineedit import CompletionLineEdit